Fixing Linux headless wallet build
The headless wallet makefile (src/makefile.unix) was broken
because it referred to files in the hash folder (src/hash/blake.c)
as being in the obj folder (src/obj/blake.c). This commit fixes
the references to these files.

A second, more serious issue was the lack of typecasting done in
the source files in src/hash. It is likely that the MinGW
gcc/g++ builds are more permissive than Debian's default gcc/g++
builds are. As such, the issue was resolved by properly casting
void pointers in each source file.